@@ -5,29 +5,28 @@ using Nethermind.Core.Extensions;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Evm;
 using Nethermind.Evm.Test;
+using Nethermind.Test.Runner;
 using NUnit.Framework;
 
-namespace Nethermind.State.Test.Runner.Test
+namespace Nethermind.State.Test.Runner.Test;
+
+public class StateTestTxTracerTest : VirtualMachineTestsBase
 {
-    [TestFixture]
-    public class StateTestTxTracerTest : VirtualMachineTestsBase
+    private StateTestTxTracer tracer;
+
+    [SetUp]
+    public void StateTestTxTracerSetUp() => tracer = new StateTestTxTracer();
+
+    [TearDown]
+    public void StateTestTxTracerTearDown() => tracer.Dispose();
+
+    [Test]
+    public void Does_not_throw_on_call()
     {
-        private StateTestTxTracer tracer;
+        byte[] code = Prepare.EvmCode
+            .CallWithValue(TestItem.AddressC, 50000, 1000000.Ether())
+            .Done;
 
-        [SetUp]
-        public void SetUp()
-        {
-            tracer = new StateTestTxTracer();
-        }
-
-        [Test]
-        public void Does_not_throw_on_call()
-        {
-            byte[] code = Prepare.EvmCode
-                .CallWithValue(TestItem.AddressC, 50000, 1000000.Ether())
-                .Done;
-
-            Assert.DoesNotThrow(() => Execute<StateTestTxTracer>(tracer, code));
-        }
+        Assert.DoesNotThrow(() => Execute(tracer, code));
     }
 }
