@@ -3,7 +3,7 @@
 
 using Newtonsoft.Json;
 
-namespace Nethermind.State.Test.Runner
+namespace Nethermind.Test.Runner
 {
     public class StateTestTxTraceResult
     {
